Refactor session fixtures to relocate for general test access
