+++ tests/unit/plugins/modules/service_config/test_service_config.py
@@ -19,183 +19,29 @@ from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
 import logging
-import os
 import pytest
-import random
-import string
 
 from pathlib import Path
-from time import sleep
 
 from cm_client import (
-    ApiClient,
-    ApiClusterList,
-    ApiCluster,
-    ApiCommand,
     ApiConfig,
-    ApiHostRef,
-    ApiHostRefList,
     ApiService,
     ApiServiceConfig,
     ApiServiceList,
     ClustersResourceApi,
-    CommandsResourceApi,
-    Configuration,
-    HostsResourceApi,
-    ParcelResourceApi,
-    ParcelsResourceApi,
     ServicesResourceApi,
 )
-from cm_client.rest import ApiException, RESTClientObject
+from cm_client.rest import ApiException
 
 from ansible_collections.cloudera.cluster.plugins.modules import service_config
-from ansible_collections.cloudera.cluster.plugins.module_utils.parcel_utils import (
-    Parcel,
-)
 
 from ansible_collections.cloudera.cluster.tests.unit import (
     AnsibleExitJson,
     AnsibleFailJson,
+    wait_for_command,
 )
 
 LOG = logging.getLogger(__name__)
-
-
-@pytest.fixture(scope="session")
-def conn():
-    conn = dict(username=os.getenv("CM_USERNAME"), password=os.getenv("CM_PASSWORD"))
-
-    if os.getenv("CM_HOST", None):
-        conn.update(host=os.getenv("CM_HOST"))
-
-    if os.getenv("CM_PORT", None):
-        conn.update(port=os.getenv("CM_PORT"))
-
-    if os.getenv("CM_ENDPOINT", None):
-        conn.update(url=os.getenv("CM_ENDPOINT"))
-
-    if os.getenv("CM_PROXY", None):
-        conn.update(proxy=os.getenv("CM_PROXY"))
-
-    return {
-        **conn,
-        "verify_tls": "no",
-        "debug": "no",
-    }
-
-
-@pytest.fixture(scope="session")
-def cm_api_client(conn):
-    """Create a Cloudera Manager API client, resolving HTTP/S and version URL.
-
-    Args:
-        conn (dict): Connection details
-
-    Returns:
-        ApiClient: Cloudera Manager API client
-    """
-    config = Configuration()
-
-    config.username = conn["username"]
-    config.password = conn["password"]
-
-    if "url" in conn:
-        config.host = str(conn["url"]).rstrip(" /")
-    else:
-        rest = RESTClientObject()
-
-        # Handle redirects
-        url = rest.GET(conn["host"]).urllib3_response.geturl()
-
-        # Get version
-        auth = config.auth_settings().get("basic")
-        version = rest.GET(
-            f"{url}api/version", headers={auth["key"]: auth["value"]}
-        ).data
-
-        # Set host
-        config.host = f"{url}api/{version}"
-
-    client = ApiClient()
-    client.user_agent = "pytest"
-    return client
-
-
-@pytest.fixture(scope="session")
-def target_cluster(cm_api_client, request):
-    """Create a 7.1.9 test cluster."""
-
-    name = (
-        Path(request.fixturename).stem
-        + "_"
-        + "".join(random.choices(string.ascii_lowercase, k=6))
-    )
-    cdh_version = "7.1.9"
-
-    cluster_api = ClustersResourceApi(cm_api_client)
-    parcels_api = ParcelsResourceApi(cm_api_client)
-    parcel_api = ParcelResourceApi(cm_api_client)
-    host_api = HostsResourceApi(cm_api_client)
-
-    try:
-        # TODO Query for the latest version available - is this possible?
-
-        # Create the initial cluster
-        config = ApiCluster(
-            name=name,
-            full_version=cdh_version,
-        )
-
-        cluster_api.create_clusters(body=ApiClusterList(items=[config]))
-
-        # Get first free host and assign to the cluster
-        all_hosts = host_api.read_hosts()
-        host = next((h for h in all_hosts.items if not h.cluster_ref), None)
-
-        if host is None:
-            # Roll back the cluster and then raise an error
-            cluster_api.delete_cluster(cluster_name=name)
-            raise Exception("No available hosts to allocate to new cluster")
-        else:
-            cluster_api.add_hosts(
-                cluster_name=name,
-                body=ApiHostRefList(items=[ApiHostRef(host_id=host.host_id)]),
-            )
-
-        # Find the first CDH parcel version and activate it
-        parcels = parcels_api.read_parcels(cluster_name=name)
-        cdh_parcel = next((p for p in parcels.items if p.product == "CDH"))
-
-        parcel = Parcel(
-            parcel_api=parcel_api,
-            product=cdh_parcel.product,
-            version=cdh_parcel.version,
-            cluster=name,
-        )
-
-        parcel.activate()
-
-        # Reread and return the cluster
-        yield cluster_api.read_cluster(cluster_name=name)
-
-        # Deprovision the cluster
-        cluster_api.delete_cluster(cluster_name=name)
-    except ApiException as ae:
-        raise Exception(str(ae))
-
-
-def wait_for_command(
-    api_client: ApiClient, command: ApiCommand, polling: int = 120, delay: int = 5
-):
-    poll_count = 0
-    while command.active:
-        if poll_count > polling:
-            raise Exception("CM command timeout")
-        sleep(delay)
-        poll_count += 1
-        command = CommandsResourceApi(api_client).read_command(command.id)
-    if not command.success:
-        raise Exception(f"CM command [{command.id}] failed: {command.result_message}")
 
 
 @pytest.fixture(scope="module")
